Fix cycle_clock data type mismatch.

+++ instructions.h
@@ -19,6 +19,8 @@
 #ifndef INCLUDED_INSTRUCTIONS_H
 #define INCLUDED_INSTRUCTIONS_H
 
+#include <stdint.h>
+
 #define F_RESET  0x1
 #define F_IRQ    0x2
 #define F_NMI    0x4
@@ -37,7 +39,7 @@ extern unsigned char A, X, Y, P, S;
 extern unsigned char N, V, B, D, I, Z, C;
 extern unsigned char interrupt_flags;
 extern unsigned short emPC;
-extern unsigned long long cycle_clock;
+extern uint64_t cycle_clock;
 
 unsigned char build_P (void);
 void unbuild_P (unsigned char status);
